# apps/memoro/apps/server/openapi.yaml
openapi: 3.1.0
info:
  title: Memoro Server API
  description: AI-powered voice memo management — memo processing, spaces, credits, meetings, settings.
  version: 1.0.0
  contact:
    name: Mana
    url: https://mana.how
servers:
  - url: http://localhost:3015
    description: Local development
  - url: https://memoro.mana.how
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceKey:
      type: apiKey
      in: header
      name: X-Service-Key
    internalKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
  schemas:
...
    CreateBotBody:
      type: object
      required: [meeting_url]
      properties:
        meeting_url:
          type: string
          pattern: '^https://(teams\.microsoft\.com|meet\.google\.com|[\w-]+\.zoom\.us)/'
        space_id:
          type: string
          format: uuid
    CheckCreditsBody:
      type: object
      required: [operation, amount]
      properties:
        operation:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0
    ConsumeCreditsBody:
      type: object
      required: [operation, amount, description]
      properties:
        operation:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0
        description:
          type: string
          minLength: 1
        metadata:
          type: object
    UpdateProfileBody:
      type: object
      properties:
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
          maxLength: 500
    UpdateDataUsageBody:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
...
paths:
...
  /api/v1/credits/balance:
    get:
      tags: [Credits]
      summary: Get credit balance
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Credit balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  credits: { type: number }
                  totalEarned: { type: number }
                  totalSpent: { type: number }
  /api/v1/credits/check:
    post:
      tags: [Credits]
      summary: Check if user has enough credits
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckCreditsBody'
      responses:
        '200': { description: Credit validation result }
  /api/v1/credits/consume:
    post:
      tags: [Credits]
      summary: Consume credits
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumeCreditsBody'
      responses:
        '200': { description: Credits consumed }

  # ── Settings ─────────────────────────────────────────────────────
  /api/v1/settings:
    get:
      tags: [Settings]
      summary: Get all user settings
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: User settings }
  /api/v1/settings/memoro:
    get:
      tags: [Settings]
      summary: Get memoro-specific settings
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Memoro settings }
    patch:
      tags: [Settings]
      summary: Update memoro settings
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
      responses:
        '200': { description: Settings updated }
        '400': { description: Empty body }
  /api/v1/settings/memoro/data-usage:
    patch:
      tags: [Settings]
      summary: Update data usage acceptance
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDataUsageBody'
      responses:
        '200': { description: Data usage updated }
  /api/v1/settings/profile:
    patch:
      tags: [Settings]
      summary: Update user profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200': { description: Profile updated }
        '400': { description: Validation error }

  # ── Meetings ─────────────────────────────────────────────────────
  /api/v1/meetings/bots:
    post:
      tags: [Meetings]
      summary: Create meeting bot
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotBody'
      responses:
        '200': { description: Bot created }
        '400': { description: Invalid meeting URL }
        '402': { description: Insufficient credits }
    get:
      tags: [Meetings]
      summary: List meeting bots
      security: [{ bearerAuth: [] }]
      parameters:
        - name: space_id
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200': { description: List of bots }
  /api/v1/meetings/bots/{id}:
    get:
      tags: [Meetings]
      summary: Get bot by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Bot details }
        '404': { description: Bot not found }
  /api/v1/meetings/bots/{id}/stop:
    post:
      tags: [Meetings]
      summary: Stop a meeting bot
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Bot stop signal sent }
        '404': { description: Bot not found }
  /api/v1/meetings/recordings:
    get:
      tags: [Meetings]
      summary: List recordings
      security: [{ bearerAuth: [] }]
      parameters:
        - name: space_id
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200': { description: List of recordings }
  /api/v1/meetings/recordings/{id}:
    get:
      tags: [Meetings]
      summary: Get recording by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Recording details }
        '404': { description: Recording not found }
  /api/v1/meetings/recordings/{id}/to-memo:
    post:
      tags: [Meetings]
      summary: Convert recording to memo
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Memo created from recording }
        '400': { description: No audio/video file }
        '404': { description: Recording not found }

  # ── Internal ─────────────────────────────────────────────────────
...